+++ src/perft.h
@@ -10,8 +10,8 @@
 struct Perft
 {
     static uint64_t perft(PositionBase& pos, const Ply depth);
-    static void perft_deepen(PositionBase& pos, const Ply depth, std::vector<uint64_t>* verfication = nullptr);
-    static void perft_deepen(const Fen& fen, const Ply depth, std::vector<uint64_t>* verfication = nullptr);
+    static void perft_deepen(PositionBase& pos, const Ply depth, std::vector<uint64_t>* verification = nullptr);
+    static void perft_deepen(const Fen& fen, const Ply depth, std::vector<uint64_t>* verification = nullptr);
     static void perft_suite();
 };